requirements.txt
# Dépendances pour le web scraping et Kafka
beautifulsoup4>=4.9.3
requests>=2.25.1
kafka-python>=2.0.2

# Dépendances pour le traitement Spark
pyspark>=3.3.0
pyarrow>=9.0.0

# Dépendances pour les analyses de données
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.1

# Dépendances pour le client API
requests>=2.25.1
python-dotenv>=0.19.0

# Dépendances pour les tests
pytest>=7.0.0
pytest-mock>=3.7.0

# Dépendances utilitaires
tqdm>=4.61.0
pyyaml>=6.0
prometheus_client